Struct ntr::Connection

pub struct Connection { /* fields omitted */ }

A connection to a 3DS.

Methods

impl Connection

fn new(addr: &str) -> Result<Self>

Opens a connection to the 3DS with the address addr.

Examples

use ntr::Connection;

let mut connection = Connection::new("192.168.2.247").expect("io error");

fn get_pid(&mut self, tid: u64) -> Result<Option<u32>>

Returns the process identifier for the currently running title id tid.

You can find a list of title ids for 3DS games at 3dsdb.

Examples

use ntr::Connection;

let pid = connection.get_pid(0x0004000000126300u64)
    .expect("io error")
    .expect("pid not found");

fn mem_read(&mut self, addr: u32, size: u32, pid: u32) -> Result<Box<[u8]>>

Reads a chunk of 3DS memory.

Reads size bytes of 3DS memory starting from address addr for the process with process id pid.

fn mem_write(&mut self, addr: u32, data: &[u8], pid: u32) -> Result<usize>

Writes data to 3DS memory.

Writes data to the 3DS memory starting at address addr for the process with process id pid.

fn read_u32(&mut self, addr: u32, pid: u32) -> Result<u32>

Reads a u32 from 3DS memory.

fn read_u16(&mut self, addr: u32, pid: u32) -> Result<u16>

Reads a u16 from 3DS memory.

fn read_u8(&mut self, addr: u32, pid: u32) -> Result<u8>

Reads a u8 from 3DS memory.

fn read_i32(&mut self, addr: u32, pid: u32) -> Result<i32>

Reads an i32 from 3DS memory.

fn read_i16(&mut self, addr: u32, pid: u32) -> Result<i16>

Reads an i16 from 3DS memory.

fn read_i8(&mut self, addr: u32, pid: u32) -> Result<i8>

Reads an i8 from 3DS memory.

fn write_u32(&mut self, addr: u32, data: u32, pid: u32) -> Result<()>

Writes a u32 to 3DS memory.

fn write_u16(&mut self, addr: u32, data: u16, pid: u32) -> Result<()>

Writes a u16 to 3DS memory.

fn write_u8(&mut self, addr: u32, data: u8, pid: u32) -> Result<()>

Writes a u8 to 3DS memory.

fn write_i32(&mut self, addr: u32, data: i32, pid: u32) -> Result<()>

Writes an i32 to 3DS memory.

fn write_i16(&mut self, addr: u32, data: i16, pid: u32) -> Result<()>

Writes an i16 to 3DS memory.

fn write_i8(&mut self, addr: u32, data: i8, pid: u32) -> Result<()>

Writes an i8 to 3DS memory.